deployment su gestionestampa.com e introdotto braintree

+++ PapiroMVC/Areas/Working/Models/Cost/Printing-Cost/ProductPartPrintingSheetGainBook.cs
@@ -66,8 +66,8 @@ namespace PapiroMVC.Models
 
             try
             {
-                int gain1_1 = (int)decimal.Truncate(LargerFormat.GetSide1() / SmallerFormat.GetSide1());
-                int gain2_2 = (int)decimal.Truncate(LargerFormat.GetSide2() / SmallerFormat.GetSide2());
+                int gain1_1 = (int)decimal.Truncate((decimal)(LargerFormat.GetSide1() / SmallerFormat.GetSide1()));
+                int gain2_2 = (int)decimal.Truncate((decimal)(LargerFormat.GetSide2() / SmallerFormat.GetSide2()));
 
                 int gain2_2Perf = (gain2_2 % 2) == 0 ? gain2_2 : gain2_2 - 1;
                 int gain1_1Perf = (gain1_1 % 2) == 0 ? gain1_1 : gain1_1 - 1;
@@ -76,8 +76,8 @@ namespace PapiroMVC.Models
                 var gSideOnSide16 = gain1_1 * gain2_2Perf;
                 var gSideOnSide12 = gain1_1Perf * gain2_2;
 
-                var gain1_2 = (int)decimal.Truncate(LargerFormat.GetSide1() / SmallerFormat.GetSide2());
-                var gain2_1 = (int)decimal.Truncate(LargerFormat.GetSide2() / SmallerFormat.GetSide1());
+                var gain1_2 = (int)decimal.Truncate((decimal)(LargerFormat.GetSide1() / SmallerFormat.GetSide2()));
+                var gain2_1 = (int)decimal.Truncate((decimal)(LargerFormat.GetSide2() / SmallerFormat.GetSide1()));
 
                 int gain2_1Perf = (gain2_1 % 2) == 0 ? gain2_1 : gain2_1 - 1;
                 int gain1_2Perf = (gain1_2 % 2) == 0 ? gain1_2 : gain1_2 - 1;
